updated branch for recent changes, and test stage selection menu
Implemented the logic to decide between different stages chosen at a seperate menu, just need the stages now to fully implement

+++ States/js/css.js
@@ -1,7 +1,12 @@
-console.log("winstate reached");
-
 var cssState={
   create: function(){
+    //Reset values to default so if player wants to play again, it does not start off "ready" to play
+    charSelected1 = false;
+    charSelected2 = false;
+    charName1 = "";
+    charName2 = "";
+
+
     key1 = game.input.keyboard.addKey(Phaser.Keyboard.ONE);
     dudeIcon = game.add.sprite(game.world.width * .5 - 200, game.world.height * .25 + 50, 'dudeIcon');
     dudeIcon.anchor.setTo(.5,.5);
@@ -14,6 +19,13 @@ var cssState={
     chickIcon.scale.setTo(.5,.5);
     game.physics.arcade.enable(chickIcon);
     chickIcon.tint =  0xffffff;
+
+    //TEST:COMPUTER icon
+    computerIcon = game.add.sprite(game.world.width * .5, game.world.height * .5  - 100, 'computerIcon');
+    computerIcon.anchor.setTo(.5,.5);
+    computerIcon.scale.setTo(.25,.25);
+    game.physics.arcade.enable(computerIcon);
+    computerIcon.tint =  0xffffff;
 
     player1Icon = game.add.sprite(game.world.width * .5 -50, game.world.height * .5, 'player1cssIcon');
     player2Icon = game.add.sprite(game.world.width * .5 +50, game.world.height * .5, 'player2cssIcon');
@@ -37,14 +49,20 @@ var cssState={
 
     dudeIcon.enableBody = true;
     chickIcon.enableBody = true;
+
+    buttonSound = game.add.audio('buttonSound');
+
     var startLabel=game.add.text(80,game.world.height-40,'Press "1" key to play game after selecting characters!',{font: '25px Arial',fill:'#ffffff'});
-    //var wkey= game.input.keyboard.addKey(Phaser.Keyboard.W);
-    //wkey.onDown.addOnce(this.start,this);
+    gameReadyText = game.add.text(game.world.width * .5,game.world.height-300,'',{font: '50px Arial',fill:'#ffffff'});
+    gameReadyText.anchor.setTo(.5,.5);
+
     player1Text = game.add.text(80,game.world.height-60,'Character 1 selected: ',{font: '25px Arial',fill:'#ffffff'});
     player2Text = game.add.text(80,game.world.height-80,'Character 2 selected: ',{font: '25px Arial',fill:'#ffffff'});
-    gameReadyText = game.add.text(80,game.world.height-100,'',{font: '25px Arial',fill:'#ffffff'});
+
     player1BodyIcon = game.add.sprite(game.world.width * .25, game.world.height * .75, '');
     player2BodyIcon = game.add.sprite(game.world.width * .75, game.world.height * .75, '');
+    player1BodyIcon.scale.setTo(1.5,1.5);
+    player2BodyIcon.scale.setTo(1.5,1.5);
 
 //TODO:Incorperate dragUpdate function event system into current system. I think it's needed to fix bugs/add dynamic features like spawning the character when hovering over while still dragging.
 //TODO:
@@ -52,9 +70,9 @@ var cssState={
 
   },
   start: function(){
+    gameReadyText.text = `Game Start!`;
     music.stop();
-
-   game.state.start('play');
+    game.state.start('sss');
  },
  update: function() {
    player1Text.text = `Character selected 1: ${charName1}`;
@@ -74,31 +92,33 @@ var cssState={
    if(charSelected1 && charSelected2 && key1.isDown)
    {
      //Eventually allow the player to start game;
-     gameReadyText.text = `Game ready`;
-     game.state.start('play');
-
-
-
-
+     gameReadyText.text = `Game Start!`;
+     game.state.start('sss');
    }
    else if(charSelected1 && charSelected2)
-   {
-     //Eventually allow the player to start game;
-     gameReadyText.text = `Game ready`;
+   { //Allow the player to tap game ready to start game
+     gameReadyText.text = `Game ready:\nClick to start!`;
+     gameReadyText.inputEnabled = true;
+     gameReadyText.events.onInputUp.addOnce(function() {
+       music.stop();
+      game.state.start('sss');
+     });
 
    }
    else {
      {
        gameReadyText.text = ``;
+        gameReadyText.inputEnabled = false;
      }
+
    }
  },
  onDragStop: function() {
 
-//If you drop the curser on the icon
+//If you drop the cursor on the icon
    if(game.physics.arcade.overlap(player1Icon, dudeIcon))
    {
-
+     buttonSound.play();
      //Determine's what's spawned, and lets you start game
      charName1 = "dude";
      charSelected1 = true;
@@ -110,7 +130,7 @@ var cssState={
      player1BodyIcon = game.add.sprite(game.world.width * .25 - 100, game.world.height * .5, 'dude');
 
      player1BodyIcon.scale.setTo(3.5,3.5);
-     player1BodyIcon.animations.add('idle', [0, 1], 5, true);
+     player1BodyIcon.animations.add('idle', [1, 2], 5, true);
      player1BodyIcon.animations.add('kick', [6], 5, true);
      if(player1BodyIcon.animations)
      {
@@ -125,6 +145,7 @@ var cssState={
    //If you drop the icon on the chick Picture
    if(game.physics.arcade.overlap(player1Icon, chickIcon))
    {
+     buttonSound.play();
      charName1 = "chick";
      charSelected1 = true;
      chickIcon.tint =  0xffff00;
@@ -133,7 +154,7 @@ var cssState={
      player1BodyIcon = game.add.sprite(game.world.width * .25 - 100, game.world.height * .5, 'chick');
 
      player1BodyIcon.scale.setTo(3.5,3.5);
-     player1BodyIcon.animations.add('idle', [0, 1], 5, true);
+     player1BodyIcon.animations.add('idle', [1, 2], 5, true);
      player1BodyIcon.animations.add('kick', [6], 5, true);
      if(player1BodyIcon.animations)
      {
@@ -145,6 +166,7 @@ var cssState={
 
    if(game.physics.arcade.overlap(player2Icon,dudeIcon))
    {
+     buttonSound.play();
      charName2 = "dude";
      charSelected2 = true;
      dudeIcon.tint =  0xffff00;
@@ -152,7 +174,7 @@ var cssState={
 
      player2BodyIcon = game.add.sprite(game.world.width * .75 - 100, game.world.height * .5, 'dude');
      player2BodyIcon.scale.setTo(3.5,3.5);
-     player2BodyIcon.animations.add('idle', [0, 1], 5, true);
+     player2BodyIcon.animations.add('idle', [1, 2], 5, true);
      player2BodyIcon.animations.add('kick', [6], 5, true);
      player2BodyIcon.visible = true;
 
@@ -171,6 +193,7 @@ var cssState={
 
    if(game.physics.arcade.overlap(player2Icon,chickIcon))
    {
+     buttonSound.play();
      charName2 = "chick";
      charSelected2 = true;
      chickIcon.tint =  0xffff00;
@@ -178,7 +201,31 @@ var cssState={
 
      player2BodyIcon = game.add.sprite(game.world.width * .75 - 100, game.world.height * .5, 'chick');
      player2BodyIcon.scale.setTo(3.5,3.5);
-     player2BodyIcon.animations.add('idle', [0, 1], 5, true);
+     player2BodyIcon.animations.add('idle', [1, 2], 5, true);
+     player2BodyIcon.animations.add('kick', [6], 5, true);
+
+     if(player2BodyIcon.animations)
+     {
+       player2BodyIcon.alpha = 1;
+     }
+   }
+   else
+   {
+    // player2BodyIcon.kill();
+   }
+
+   if(game.physics.arcade.overlap(player2Icon,computerIcon))
+   {
+     buttonSound.play();
+     charName2 = "chick";
+     charSelected2 = true;
+     computerIcon.tint =  0xffff00;
+     player2BodyIcon.kill();
+     controlOptionAI = -2; //Temporary till we have the AI logic, then replace this with a -2 instead,using vpad to test functionality
+     console.log("controlOptionAI: " + controlOptionAI);
+     player2BodyIcon = game.add.sprite(game.world.width * .75 - 100, game.world.height * .5, 'chick');
+     player2BodyIcon.scale.setTo(3.5,3.5);
+     player2BodyIcon.animations.add('idle', [1, 2], 5, true);
      player2BodyIcon.animations.add('kick', [6], 5, true);
 
      if(player2BodyIcon.animations)
@@ -196,7 +243,7 @@ var cssState={
      player1BodyIcon.kill();
    }
 
-   if(!game.physics.arcade.overlap(player2Icon,dudeIcon) && !game.physics.arcade.overlap(player2Icon,chickIcon))
+   if(!game.physics.arcade.overlap(player2Icon,dudeIcon) && !game.physics.arcade.overlap(player2Icon,chickIcon) && !game.physics.arcade.overlap(player2Icon,computerIcon))
    {
      player2BodyIcon.kill();
    }
@@ -207,8 +254,8 @@ var cssState={
 
    if(game.physics.arcade.overlap(player1Icon, dudeIcon))
    {
-     charName2 = "";
-     charSelected2 = false;
+     charName1 = "";
+     charSelected1 = false;
      dudeIcon.tint =  0xffffff;
 
       if(player1BodyIcon.animations)
@@ -249,6 +296,19 @@ var cssState={
      charName2 = "";
      charSelected2 = false;
      chickIcon.tint =  0xffffff;
+
+      if(player2BodyIcon.animations)
+      {
+        player2BodyIcon.alpha = .5;
+      }
+   }
+   if(game.physics.arcade.overlap(player2Icon, computerIcon))
+   {
+     charName2 = "";
+     charSelected2 = false;
+     computerIcon.tint =  0xffffff;
+     controlOptionAI = 2;
+     console.log("controlOptionAI: " + controlOptionAI);
 
       if(player2BodyIcon.animations)
       {
